test(cli): migrate cli tests to vitest
