+++ packages/@sanity/cli/package.json
@@ -51,7 +51,7 @@
     "clean": "rimraf lib",
     "lint": "eslint .",
     "prepublishOnly": "turbo run build",
-    "test": "jest",
+    "test": "vitest",
     "ts": "node -r esbuild-register",
     "watch": "pkg-utils watch"
   },
@@ -75,7 +75,6 @@
     "validate-npm-package-name": "^3.0.0"
   },
   "devDependencies": {
-    "@jest/globals": "^29.7.0",
     "@repo/package.config": "workspace:*",
     "@repo/test-config": "workspace:*",
     "@rexxars/gitconfiglocal": "^3.0.1",
