+++ packages/@sanity/cli/test/basics.test.ts
@@ -1,6 +1,6 @@
 import path from 'node:path'
 
-import {describe, expect} from '@jest/globals'
+import {describe, expect} from 'vitest'
 
 import {describeCliTest, testConcurrent} from './shared/describe'
 import {getCliUserEmail, runSanityCmdCommand, studioVersions} from './shared/environment'
