+++ packages/@sanity/cli/test/dev.test.ts
@@ -2,7 +2,7 @@ import {createHash} from 'node:crypto'
 import {readFile} from 'node:fs/promises'
 import path from 'node:path'
 
-import {describe, expect, test} from '@jest/globals'
+import {describe, expect, test} from 'vitest'
 
 import {describeCliTest} from './shared/describe'
 import {testServerCommand} from './shared/devServer'
