+++ packages/@sanity/cli/test/init.test.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs/promises'
 import path from 'node:path'
 
-import {describe, expect} from '@jest/globals'
+import {describe, expect} from 'vitest'
 
 import templates from '../src/actions/init-project/templates'
 import {describeCliTest, testConcurrent} from './shared/describe'
