+++ packages/@sanity/cli/test/shared/describe.ts
@@ -1,4 +1,4 @@
-import {describe, test} from '@jest/globals'
+import {describe, test} from 'vitest'
 
 import {cliUserToken, hasBuiltCli} from './environment'
