+++ packages/@sanity/cli/test/typegen.test.ts
@@ -1,6 +1,6 @@
 import {readFile, writeFile} from 'node:fs/promises'
 
-import {expect, test} from '@jest/globals'
+import {expect, test} from 'vitest'
 
 import {describeCliTest} from './shared/describe'
 import {runSanityCmdCommand, studiosPath} from './shared/environment'
