+++ packages/@sanity/cli/vitest.config.mts
@@ -1,0 +1,9 @@
+import {defineConfig} from '@repo/test-config/vitest'
+
+export default defineConfig({
+  test: {
+    includeSource: ['./test/**/*.ts'],
+    globalSetup: ['./test/shared/globalSetup.ts'],
+    testTimeout: 30000, // 30s
+  },
+})
